updates entities - ovverirde equals and hashCode
