+++ src/main/java/com/cova/quizapp/data/TriviaHistoryRepo.java
@@ -1,6 +1,6 @@
 package com.cova.quizapp.data;
 
-import com.cova.quizapp.model.persistence.TriviaHistory;
+import com.cova.quizapp.model.entity.TriviaHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
