+++ src/main/java/com/cova/quizapp/data/TriviaRepository.java
@@ -1,6 +1,6 @@
 package com.cova.quizapp.data;
 
-import com.cova.quizapp.model.persistence.Trivia;
+import com.cova.quizapp.model.entity.Trivia;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
