+++ src/main/java/com/cova/quizapp/data/UserRepository.java
@@ -1,6 +1,6 @@
 package com.cova.quizapp.data;
 
-import com.cova.quizapp.model.persistence.AppUser;
+import com.cova.quizapp.model.entity.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
