+++ src/main/java/com/cova/quizapp/model/entity/AppUser.java
@@ -1,4 +1,4 @@
-package com.cova.quizapp.model.persistence;
+package com.cova.quizapp.model.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
@@ -55,6 +55,33 @@ public class AppUser {
     @OneToMany(mappedBy = "appUser", fetch = FetchType.LAZY)
     @JsonIgnore
     private List<TriviaHistory> triviaHistories;
+
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((id == null) ? 0 : id.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        AppUser other = (AppUser) obj;
+        if (id == null) {
+            if (other.id != null)
+                return false;
+        } else if (!id.equals(other.id))
+            return false;
+        return true;
+    }
+
 
 
 }
