+++ src/main/java/com/cova/quizapp/model/entity/Trivia.java
@@ -1,0 +1,62 @@
+package com.cova.quizapp.model.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "trivia")
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Trivia {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id ;
+
+    @Column(name = "question")
+    private String question;
+
+    @Column(name = "answer")
+    private String answer;
+
+    @Column(name = "difficulty_level")
+    @Enumerated(EnumType.STRING)
+    private DifficultyLevel difficultyLevel;
+
+    public enum DifficultyLevel{
+        HARD,
+        MEDIUM,
+        EASY
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((id == null) ? 0 : id.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        Trivia other = (Trivia) obj;
+        if (id == null) {
+            if (other.id != null)
+                return false;
+        } else if (!id.equals(other.id))
+            return false;
+        return true;
+    }
+
+
+}
