--- src/main/java/com/cova/quizapp/model/persistence/Trivia.java
@@ -1,37 +1,0 @@
-package com.cova.quizapp.model.persistence;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "trivia")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Trivia {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id ;
-
-    @Column(name = "question")
-    private String question;
-
-    @Column(name = "answer")
-    private String answer;
-
-    @Column(name = "difficulty_level")
-    @Enumerated(EnumType.STRING)
-    private DifficultyLevel difficultyLevel;
-
-    public enum DifficultyLevel{
-        HARD,
-        MEDIUM,
-        EASY
-    }
-
-
-}
