+++ src/main/java/com/cova/quizapp/serviceimpl/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
 package com.cova.quizapp.serviceimpl;
 
 import com.cova.quizapp.data.UserRepository;
-import com.cova.quizapp.model.persistence.AppUser;
+import com.cova.quizapp.model.entity.AppUser;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,7 +9,6 @@ import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
